LayoutTests/css3/calc/font-size-expected.txt

Tests that CSS3 calc() can be used with the font-size property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getComputedStyle(document.getElementById("calc-percent"), null).fontSize should be 40px. Was 20px.
FAIL getComputedStyle(document.getElementById("calc-percent-pixels"), null).fontSize should be 40px. Was 20px.
PASS successfullyParsed is true

TEST COMPLETE
